requirements.txt
numpy<2
pandas>=1.5.3,<2
pyarrow>=10.0.1,<12
google-cloud-bigquery>=3.12.0,<4
selenium>=4.8.0,<5
beautifulsoup4>=4.12.2,<5
chromedriver-autoinstaller>=0.4.0,<1
python-dotenv>=1.0.0,<2
pandas-gbq>=0.26.1,<1
scikit-learn>=1.2.0,<2
joblib>=1.2.0,<2